chore: 필요한 주석추가, 필요없는 내용 삭제
